Enable build warnings/errors on Windows (#831)
* Enable build warnings/errors on Windows

src/h3lib/lib/coodijk.c:

    (73,10): warning C4244: '=': conversion from 'double' to 'int', possible loss of data
    (74,10): warning C4244: '=': conversion from 'double' to 'int', possible loss of data
    (133,25): warning C4244: '=': conversion from 'double' to 'int', possible loss of data
    (137,25): warning C4244: '=': conversion from 'double' to 'int', possible loss of data
    (236,13): warning C4307: '+': signed integral constant overflow

src/h3lib/lib/localij.c:

    (633,14): warning C4244: 'initializing': conversion from 'double' to 'int', possible loss of data
    (634,14): warning C4244: 'initializing': conversion from 'double' to 'int', possible loss of data
    (635,14): warning C4244: 'initializing': conversion from 'double' to 'int', possible loss of data

src/apps/testapps/testH3Memory.c:

    (84,21): warning C4098: 'test_prefix_free': 'void' function returning a value

src/apps/testapps/testMathExtensionsInternal.c:

    (81,9): warning C4307: '-': signed integral constant overflow
    (82,9): warning C4307: '-': signed integral constant overflow
    (83,9): warning C4307: '-': signed integral constant overflow
    (84,9): warning C4307: '-': signed integral constant overflow
    (85,9): warning C4307: '-': signed integral constant overflow
    (86,9): warning C4307: '-': signed integral constant overflow
    (87,9): warning C4307: '-': signed integral constant overflow
    (88,9): warning C4307: '-': signed integral constant overflow
    (89,9): warning C4307: '-': signed integral constant overflow
    (90,9): warning C4307: '-': signed integral constant overflow
    (91,9): warning C4307: '-': signed integral constant overflow
    (92,9): warning C4307: '-': signed integral constant overflow
    (93,9): warning C4307: '-': signed integral constant overflow
    (94,9): warning C4307: '-': signed integral constant overflow
    (95,9): warning C4307: '-': signed integral constant overflow
    (96,9): warning C4307: '-': signed integral constant overflow
    (97,9): warning C4307: '-': signed integral constant overflow
    (98,9): warning C4307: '-': signed integral constant overflow
    (99,9): warning C4307: '-': signed integral constant overflow
    (100,9): warning C4307: '-': signed integral constant overflow
    (101,9): warning C4307: '-': signed integral constant overflow
    (102,9): warning C4307: '-': signed integral constant overflow
    (103,9): warning C4307: '-': signed integral constant overflow
    (104,9): warning C4307: '-': signed integral constant overflow
    (105,9): warning C4307: '-': signed integral constant overflow
    (106,9): warning C4307: '-': signed integral constant overflow

src/apps/testapps/testCellToLocalIj.c:

    (204,44): warning C4146: unary minus operator applied to unsigned type, result still unsigned

* Add ENABLE_WARNINGS option


---------

+++ src/h3lib/lib/coordijk.c
@@ -70,8 +70,8 @@ void _hex2dToCoordIJK(const Vec2d *v, CoordIJK *h) {
     x1 = a1 + x2 / 2.0;
 
     // check if we have the center of a hex
-    m1 = x1;
-    m2 = x2;
+    m1 = (int)x1;
+    m2 = (int)x2;
 
     // otherwise round correctly
     r1 = x1 - m1;
@@ -130,11 +130,11 @@ void _hex2dToCoordIJK(const Vec2d *v, CoordIJK *h) {
         {
             long long int axisi = h->j / 2;
             long long int diff = h->i - axisi;
-            h->i = h->i - 2.0 * diff;
+            h->i = (int)(h->i - 2.0 * diff);
         } else {
             long long int axisi = (h->j + 1) / 2;
             long long int diff = h->i - axisi;
-            h->i = h->i - (2.0 * diff + 1);
+            h->i = (int)(h->i - (2.0 * diff + 1));
         }
     }
